last minute fixes
